Separate StageInfo-related APIs from game-stage code
