+++ src/menu/stagepractice.c
@@ -13,6 +13,7 @@
 #include "options.h"
 #include "global.h"
 #include "video.h"
+#include "stageinfo.h"
 
 static void draw_stgpract_menu(MenuData *m) {
 	draw_options_menu_bg(m);
@@ -30,12 +31,15 @@ MenuData* create_stgpract_menu(Difficulty diff) {
 	m->flags = MF_Abortable;
 	m->transition = TransFadeBlack;
 
-	dynarray_foreach_elem(&stages, StageInfo *stg, {
+	int n = stageinfo_get_num_stages();
+	for(int i = 0; i < n; ++i) {
+		StageInfo *stg = stageinfo_get_by_index(i);
+
 		if(stg->type != STAGE_STORY) {
 			break;
 		}
 
-		StageProgress *p = stage_get_progress_from_info(stg, diff, false);
+		StageProgress *p = stageinfo_get_progress(stg, diff, false);
 
 		if(p && p->unlocked) {
 			snprintf(title, sizeof(title), "%s: %s", stg->title, stg->subtitle);
@@ -44,7 +48,7 @@ MenuData* create_stgpract_menu(Difficulty diff) {
 			snprintf(title, sizeof(title), "%s: ???????", stg->title);
 			add_menu_entry(m, title, NULL, NULL);
 		}
-	});
+	}
 
 	add_menu_separator(m);
 	add_menu_entry(m, "Back", menu_action_close, NULL);
