+++ src/stages/corotest.h
@@ -11,7 +11,7 @@
 
 #include "taisei.h"
 
-#include "stage.h"
+#include "stageinfo.h"
 
 extern StageProcs corotest_procs;
